+++ sdk/geofence/iOS/MoEReactNativeGeofenceUtil.h
@@ -1,0 +1,13 @@
+//
+//  MoEReactNativeInboxUtil.h
+//  ReactNativeMoEngageInbox
+//
+//  Created by Rakshitha . on 26/12/24.
+//
+
+
+@import MoEngagePluginGeofence;
+
+@interface MoEReactNativeGeofenceUtil : NSObject
++(NSDictionary*)getJSONRepresentation:(NSString*)string;
+@end