+++ sdk/geofence/CHANGELOG.md
@@ -1,3 +1,10 @@
+# Next release
+
+## Next version
+
+- iOS
+    - Fixed the import issue seen in React version 0.76
+    
 # 25-11-2024
 
 ## 6.0.2